# Code/gamesUP/docs/openapi.yaml
openapi: 3.0.1
info:
  title: GamesUP API
  description: |
    API REST du catalogue et des parcours clients GamesUP.
    Les ressources personnelles sont toujours déterminées par le jeton JWT :
    aucune route ne permet de sélectionner un autre utilisateur.
  contact:
    name: GamesUP
  license:
    name: Projet pédagogique
  version: 1.0.0
servers:
- url: /
  description: Serveur courant
...
paths:
...
  /api/v1/admin/inventory/{gameId}:
    get:
      tags:
      - Inventory administration
      summary: Consulter le stock d'un jeu
      operationId: find
      parameters:
      - name: gameId
        in: path
        required: true
        schema:
          type: integer
          format: int64
      responses:
        "200":
          description: OK
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/InventoryResponse"
        "400":
          description: Requête ou paramètre invalide.
          content:
            application/problem+json:
              schema:
                $ref: "#/components/schemas/ProblemDetail"
        "401":
          description: "JWT absent, invalide ou expiré."
          content:
            application/problem+json:
              schema:
                $ref: "#/components/schemas/ProblemDetail"
        "403":
          description: Rôle ou propriété de la ressource insuffisant.
          content:
            application/problem+json:
              schema:
                $ref: "#/components/schemas/ProblemDetail"
        "404":
          description: Ressource introuvable.
          content:
            application/problem+json:
              schema:
                $ref: "#/components/schemas/ProblemDetail"
      security:
      - bearerAuth: []
    put:
      tags:
      - Inventory administration
      summary: Définir le stock absolu
      operationId: setAbsoluteQuantity
      parameters:
      - name: gameId
        in: path
        required: true
        schema:
          type: integer
          format: int64
      requestBody:
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/InventoryUpdateRequest"
        required: true
      responses:
        "200":
          description: OK
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/InventoryResponse"
        "400":
          description: Requête ou paramètre invalide.
          content:
            application/problem+json:
              schema:
                $ref: "#/components/schemas/ProblemDetail"
        "401":
          description: "JWT absent, invalide ou expiré."
          content:
            application/problem+json:
              schema:
                $ref: "#/components/schemas/ProblemDetail"
        "403":
          description: Rôle ou propriété de la ressource insuffisant.
          content:
            application/problem+json:
              schema:
                $ref: "#/components/schemas/ProblemDetail"
        "404":
          description: Ressource introuvable.
          content:
            application/problem+json:
              schema:
                $ref: "#/components/schemas/ProblemDetail"
      security:
      - bearerAuth: []
...
  /api/v1/admin/orders/{orderId}/status:
    patch:
      tags:
      - Order administration
      summary: Faire évoluer le statut
      description: Seules les transitions définies par la machine d'état sont acceptées.
      operationId: transition
      parameters:
      - name: orderId
        in: path
        required: true
        schema:
          type: integer
          format: int64
      requestBody:
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/AdminOrderStatusRequest"
        required: true
      responses:
        "200":
          description: OK
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/OrderResponse"
        "400":
          description: Requête ou paramètre invalide.
          content:
            application/problem+json:
              schema:
                $ref: "#/components/schemas/ProblemDetail"
        "401":
          description: "JWT absent, invalide ou expiré."
          content:
            application/problem+json:
              schema:
                $ref: "#/components/schemas/ProblemDetail"
        "403":
          description: Rôle ou propriété de la ressource insuffisant.
          content:
            application/problem+json:
              schema:
                $ref: "#/components/schemas/ProblemDetail"
        "404":
          description: Ressource introuvable.
          content:
            application/problem+json:
              schema:
                $ref: "#/components/schemas/ProblemDetail"
        "422":
          description: Règle métier empêchant l'opération.
          content:
            application/problem+json:
              schema:
                $ref: "#/components/schemas/ProblemDetail"
      security:
      - bearerAuth: []
...
  /api/v1/admin/orders:
    get:
      tags:
      - Order administration
      summary: Lister toutes les commandes
      operationId: findAll_1
      parameters:
      - name: page
        in: query
        required: false
        schema:
          minimum: 0
          type: integer
          format: int32
          default: 0
      - name: size
        in: query
        required: false
        schema:
          maximum: 100
          minimum: 1
          type: integer
          format: int32
          default: 20
      responses:
        "200":
          description: OK
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/PageResponseOrderResponse"
        "400":
          description: Requête ou paramètre invalide.
          content:
            application/problem+json:
              schema:
                $ref: "#/components/schemas/ProblemDetail"
        "401":
          description: "JWT absent, invalide ou expiré."
          content:
            application/problem+json:
              schema:
                $ref: "#/components/schemas/ProblemDetail"
        "403":
          description: Rôle ou propriété de la ressource insuffisant.
          content:
            application/problem+json:
              schema:
                $ref: "#/components/schemas/ProblemDetail"
        "404":
          description: Ressource introuvable.
          content:
            application/problem+json:
              schema:
                $ref: "#/components/schemas/ProblemDetail"
        "422":
          description: Règle métier empêchant l'opération.
          content:
            application/problem+json:
              schema:
                $ref: "#/components/schemas/ProblemDetail"
      security:
      - bearerAuth: []
components:
  schemas:
...
    InventoryUpdateRequest:
      required:
      - quantity
      type: object
      properties:
        quantity:
          minimum: 0
          type: integer
          format: int32
    InventoryResponse:
      type: object
      properties:
        id:
          type: integer
          format: int64
        gameId:
          type: integer
          format: int64
        quantity:
          type: integer
          format: int32
        version:
          type: integer
          format: int64
...
    OrderLineResponse:
      type: object
      properties:
        gameId:
          type: integer
          format: int64
        gameName:
          type: string
        quantity:
          type: integer
          format: int32
        unitPrice:
          type: number
        lineTotal:
          type: number
    OrderResponse:
      type: object
      properties:
        id:
          type: integer
          format: int64
        customerId:
          type: integer
          format: int64
        status:
          type: string
          enum:
          - PENDING
          - PAID
          - SHIPPED
          - DELIVERED
          - CANCELLED
          - ARCHIVED
        total:
          type: number
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
        lines:
          type: array
          items:
            $ref: "#/components/schemas/OrderLineResponse"
...
    ProblemDetail:
      type: object
      properties:
        type:
          type: string
          format: uri
        title:
          type: string
        status:
          type: integer
          format: int32
        detail:
          type: string
        instance:
          type: string
          format: uri
        properties:
          type: object
          additionalProperties:
            type: object
...
    AdminOrderStatusRequest:
      required:
      - status
      type: object
      properties:
        status:
          type: string
          enum:
          - PENDING
          - PAID
          - SHIPPED
          - DELIVERED
          - CANCELLED
          - ARCHIVED
    PageResponseOrderResponse:
      type: object
      properties:
        content:
          type: array
          items:
            $ref: "#/components/schemas/OrderResponse"
        page:
          type: integer
          format: int32
        size:
          type: integer
          format: int32
        totalElements:
          type: integer
          format: int64
        totalPages:
          type: integer
          format: int32
...
  securitySchemes:
    bearerAuth:
      type: http
      description: Jeton retourné par l'inscription ou la connexion.
      scheme: bearer
      bearerFormat: JWT